- Added Backup&Restore and Export to Excel functions in FrmUserSetting.cs - Added FrmResetPassword.cs - Added UserValidator.cs - Added IssueBookValidator.cs - Added DialogSelectBook.cs - Added DialogSelectStudent.cs - Added FrmSplashScreen.cs - Logging when user login and logout

+++ LibraryDBMS/Forms/FrmManageUser.cs
@@ -31,7 +31,7 @@ namespace LibraryDBMS.Forms
             LibModule.FillDataGrid("viewUserInfo", dgvUserList, "userID");
             lblUserCount.Text = "Total User: " + 
                 LibModule.ExecuteScalarQuery("SELECT COUNT(userID) FROM viewUserInfo;");
-
+            lblRowsCount.Text = $"Total Result: {dgvUserList.Rows.Count}";
             btnEdit.Enabled = false;
             btnDelete.Enabled = false;
             btnActivateDeactivate.Enabled = false;
@@ -57,6 +57,7 @@ namespace LibraryDBMS.Forms
                                 LibModule.SearchAndFillDataGrid("viewUserInfo", "username", value, dgvUserList);
                             else if (searchBy == "Name")
                                 LibModule.SearchNameAndFillDataGrid("viewUserInfo", value, dgvUserList);
+                            lblRowsCount.Text = $"Total Result: {dgvUserList.Rows.Count}";
                         }
                     }
                     catch (Exception ex)
@@ -72,6 +73,7 @@ namespace LibraryDBMS.Forms
                             string fromDate = dtpFromDate.Value.ToString("yyyy-MM-dd");
                             string toDate = dtpToDate.Value.ToString("yyyy-MM-dd");
                             LibModule.SearchBetweenDateAndFillDataGrid("viewUserInfo", dgvUserList, "dateAdded", fromDate, toDate);
+                            lblRowsCount.Text = $"Total Result: {dgvUserList.Rows.Count}";
                         }
                     }
                     catch (Exception ex)
@@ -93,10 +95,15 @@ namespace LibraryDBMS.Forms
                 case "btnEdit":
                     try
                     {
-                        string id = dgvUserList.SelectedRows[0].Cells["userID"].Value.ToString();
-                        Form frmAddEditUser =
-                            new DialogAddEditUser(this, LibModule.GetSingleRecordFromDB("viewUserInfo", "userID", id));
-                        frmAddEditUser.ShowDialog();
+                        string userID = dgvUserList.SelectedRows[0].Cells["userID"].Value.ToString();
+
+                        if (userID != "1")
+                        {
+                            Form frmAddEditUser =
+                                                new DialogAddEditUser(this, LibModule.GetSingleRecordFromDB("viewUserInfo", "userID", userID));
+                            frmAddEditUser.ShowDialog(); 
+                        }
+                        else MessageBox.Show("Cannot edit default admin account!", "Default Admin Account", MessageBoxButtons.OK, MessageBoxIcon.Error);
                     }
                     catch (Exception ex)
                     {
@@ -107,9 +114,15 @@ namespace LibraryDBMS.Forms
                     try
                     {
                         string userID = dgvUserList.SelectedRows[0].Cells["userID"].Value.ToString();
-                        if(LibModule.DeleteRecord("tblUser", "userID", userID,
+
+                        // userID 1 = default admin account
+                        if (userID != "1")
+                        {
+                            if (LibModule.DeleteRecord("tblUser", "userID", userID,
                             Utils.GetDataGridSelectedRowData(dgvUserList)) == true)
-                            PopulateDataGrid();
+                                PopulateDataGrid();
+                        }
+                        else MessageBox.Show("Cannot delete default admin account!", "Default Admin Account", MessageBoxButtons.OK, MessageBoxIcon.Error);
                     }
                     catch (Exception ex)
                     {
@@ -120,18 +133,22 @@ namespace LibraryDBMS.Forms
                     try
                     {
                         string userID = dgvUserList.SelectedRows[0].Cells["userID"].Value.ToString();
-                        string isActive =
-                            dgvUserList.SelectedRows[0].Cells["isActive"].Value.ToString() == "Yes" ?
-                            "No" : "Yes";
-                        List<string> value = new List<string>{ isActive };
-                        if (LibModule.UpdateRecord("tblUser", "isActive", "userID", userID, value, 
-                            false) == true)
+                        if (userID != "1")
                         {
-                            MessageBox.Show($"{dgvUserList.SelectedRows[0].Cells["username"].Value}'s " +
-                                $"active status has changed!", "Active Status Updated", MessageBoxButtons.OK,
-                                MessageBoxIcon.Information);
+                            string isActive =
+                                dgvUserList.SelectedRows[0].Cells["isActive"].Value.ToString() == "Yes" ?
+                                "No" : "Yes";
+                            List<string> value = new List<string> { isActive };
+                            if (LibModule.UpdateRecord("tblUser", "isActive", "userID", userID, value,
+                                false) == true)
+                            {
+                                MessageBox.Show($"{dgvUserList.SelectedRows[0].Cells["username"].Value}'s " +
+                                    $"active status has changed!", "Active Status Updated", MessageBoxButtons.OK,
+                                    MessageBoxIcon.Information);
+                            }
+                            PopulateDataGrid(); 
                         }
-                        PopulateDataGrid();
+                        else MessageBox.Show("Cannot deactivate default admin account!", "Default Admin Account", MessageBoxButtons.OK, MessageBoxIcon.Error);
                     }
                     catch (Exception ex)
                     {
